requirements.txt
pycryptodomex
colorama
